% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myttest.R
\name{myttest}
\alias{myttest}
\title{Myttest function will investigate the data for evidence for a difference in the population variances
if the data supports the conclusion that the underlying variances are likely different or
if paired = TRUE then an appropriate t-test is made to produce a list containing:}
\usage{
myttest(x, y, alpha, paired)
}
\arguments{
\item{x}{first vector}

\item{y}{second vector}

\item{alpha}{confidence interval}

\item{paired}{experimental unit measurement}
}
\value{
list of the data frame
}
\description{
- "Welch", "T-test", "Paired" depending on what test was conducted
- A "Y/N" conclusion as to whether we should reject the NULL of equality of means at the alpha level of significance
- The summary statistics made by the appropriate t.test in a data list
}
\examples{
 myttest(x = 1:10 ,y = 1:10 ,alpha=0.05,paired=FALSE)
}
