% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.Rttest}
\alias{plot.Rttest}
\title{An informative plot with the "Rttest" class that creates boxplots of
the data when non - paired data is analyzed. It also will make a boxplot of
differences in data when paired = TRUE The ci for the difference of means
is plotted inside the boxplot of differences.}
\usage{
\method{plot}{Rttest}(x, ...)
}
\arguments{
\item{x}{named list of myttest function}

\item{...}{from ggplot function}
}
\value{
plot of the data frame
}
\description{
An informative plot with the "Rttest" class that creates boxplots of
the data when non - paired data is analyzed. It also will make a boxplot of
differences in data when paired = TRUE The ci for the difference of means
is plotted inside the boxplot of differences.
}
