% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.Rttest}
\alias{print.Rttest}
\title{Print function prints to the command line  A (1-alpha)100% confidence
interval for mu_x - mu_y and the type of test}
\usage{
\method{print}{Rttest}(x, ...)
}
\arguments{
\item{x}{named list of myttest function}

\item{...}{from print}
}
\value{
the confidence interval and type of test
}
\description{
Print function prints to the command line  A (1-alpha)100% confidence
interval for mu_x - mu_y and the type of test
}
