% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyplots.R
\name{shinyplots}
\alias{shinyplots}
\title{Dynamic plot using shiny}
\usage{
shinyplots()
}
\value{
plots and print output of package
}
\description{
Wigdets, plotting areas and layouts
}
\examples{
\dontrun{ shinyplots()}
}
